fix listenable get() performance, move out colors
